Crackling Results
Originally ran with a prefiltered 9 guides for CXCL10

took weakening the offtarget a fair bit to get any survivors.  Good for knockout, but may not be the best for repress.

Widened the search, using the wide TSS, and let Crackling have it's head in terms of the filtering.

relaxed_avg_65_wide is the result of that